+++ ios/history.md
@@ -6,6 +6,7 @@
 * Removed deprecated code to to support keyman:// scheme for ad-hoc distribution (#1160)
 * Updated the default keyboard to SIL EuroLatin (#1288)
 * Added support for error feeedback from keyboards (#1314)
+* Fixed bug behind some crashes of system keyboard (#1166)
 
 ## 2018-08-02 10.0.208 stable 
 * Fixed OSK layout problems (and possible crash) on iOS 11 on certain hardware (#1089, #1159)